chore: rename NewTable.svelte to Table.svelte
